test string/to_integer

+++ lib/string/to_integer.hpp
@@ -15,7 +15,7 @@ template<> long long toInteger<long long>(const string& s) {
   return stoll(s);
 }
 
-template<typename T = long long> T toInteger(const string& s, int n) {
+template<typename T = int64_t> T toInteger(const string& s, int n) {
   T res = 0;
   for (char c : s) {
     if (isdigit(c)) res = res * n + c - '0';
